+++ shims.js
@@ -1,0 +1,6 @@
+globalThis.CSS = { escape: input => input };
+globalThis.MutationObserver = class MutationObserver {
+	observer() {
+		//
+	}
+};